+++ cmd/influxd/upgrade/upgrade.go
@@ -171,7 +171,7 @@ func NewCommand() *cobra.Command {
 			DestP:   &options.target.configsPath,
 			Flag:    "influx-configs-path",
 			Default: filepath.Join(v2dir, "configs"),
-			Desc:    "path for CLI configurations",
+			Desc:    "path for 2.x CLI configurations file",
 			Short:   'c',
 		},
 		{